Fixed JSON serialization error
- small refactoring
  : moved get_inference_results to inference.py
  : deleted main.py
- change data type of likes numpy.int to int
